Add (incomplete) support for `omp declare target`
With this commit, sotoc now copies all declarations within an `#pragma
omp declare target` constructs into the target source code. Function
bodies which are outside an `declare target` construct are also added if
the function was declared within a `declare target` construct AND the
declaration in an `declare target` construct comes BEFORE the function's
body.