[ofld-tool-wrapper] Call .str() on stringstream before passing to cout
Empirically GCC and Intel Compiler cast to void* and output a pointer
address which is not helpful. Clang is somewhat "better" and prints
tons of error messages during compilation because libc++ in C++11 mode
doesn't allow the implicit conversion.
Long story short: Call .str() to construct a std::string and make all
compilers as well as the developers happy.