Fix Issue #20 and pretty printing decls
This adds a fix for copying decls that are declared by macros, similar
to the fix for issue #19.
It also re-introduces pretty printing of decls instead of using buffer
copying. To this end, it adds two hacks to use the correct printing
policy to not have types auto-expanded (this kinda relies on possibly
unintended clang behaviour) and to remove '#pragma omp declare target'
from the output of Decl->print().