Remove CMAKE option NECAURORA_USE_TOOLCHAIN_WRAPPERS
and default to "ON".
In was possible to set symlinks to the offload compiler driver for
assembling / linking of offload code instead of using the toolchain
wrappers.
Noone ever used this instead of the toolchain wrapper and we now support
static linking of the target code which requires the use of another
wrapper for the linker execution.