Fix issue of wrong function name for target regions
instead of the last function decl visited, the FindTargetCodeVisitor now
uses a stack of all function decls visited to determine the function in
which the target region is contained.
This fixes the issue that the visitor remembers the wrong function if
there is a function declaration before the target region but inside the
function in which the region is contained.