Remove libveoshim from the NEC SX-Aurora plugin
Libveoshim was used to basically emulate NEC's libveo functionality on
the host machine, using libffi to try to run the offloading code (which
then had to be built for the host architecture instead of the
SX-Aurora).
We had actual SX-Aurora device to test the project for some time now,
and haven't further developed or used libveoshim. This library was
always a pretty ugly hack which we can now luckily work without.
It seems unlikely that we will ever need it again so this removes it
from the project.