[LDA] Remove isSimpleAddRec() check for now
While checking that a SCEV subscript is within what we can handle
is good and probably I'll restore it at some point, it's very
restrictive. Because eventually, we'll do minusSCEV between
two subscripts. And those SCEVs can be as complicated as they want.
We don't care since we check if the result is a scConstant anyway.