[LDA] Remove unneeded doc and the expansion of dimensions
Given that we have the concept of squashing, expanding dimensions
is unneeded if we're only analyzing accesses that are supposed
to always have the same number of subscripts in the source
program (e.g. If you have int A[n][m], you can't have
an access like A[a][b] and then an access like A[n]).

The only thing we have to consider here is when the user
manually constructs multi-dimensional accesses, e.g.
A[i*m + j]. In that case we can have another access
like A[i]. Notice though that:

a) This is probably very rare.
b) Right now we can't handle it anyway even with expansion
of dimensions. This is because, it would basically result
in two accesses like A[i][j] and A[i][0]. Since the
subtraction of the second subscript, j - 0, is not
constant, we bail.