[LDA] Definitely vectorizable if distance of outer index is not 0
If two subscripts that uses outer induction variable have distance
that is not 0, then in this loop the two accesses never alias.
For example, in this:
for (i)
  for (j)
    A[i+1][j+1] = A[i][j];

The j-loop is vectorizable because the two accesses always
access different rows (the one is i, the other is i+1) so they
never alias.

This of course is yet another place that assumes that the subscripts
are in-bounds.