[VE] Use LEGALAVL nodes to annotate pack-legal VL values
Before, SDNodes were tagged as pack-legalized by tracking them in a std
of SDNodes. This is fragile, as SDNodes objects may be reused even in
the same SelectionDAG.

This new implementation wraps the VL in a 'LEGALAVL' SDNode to indicate
that it refers to packs of 64bit elements. This is more robust and will
be the basis for upstreaming.